+++ packages/vsce-api/src/interfaces/resources/index.ts
@@ -25,3 +25,5 @@ export * from "./IURIMap";
 export * from "./IWebService";
 export * from "./IJVMServer";
 export * from "./IJVMEndpoint";
+export * from "./ITSQueue";
+export * from "./ISharedTSQueue";