+++ packages/vsce-api/src/resources/ResourceTypes.ts
@@ -22,6 +22,8 @@ export enum ResourceTypes {
   CICSWebService = "CICSWebService",
   CICSJVMServer = "CICSJVMServer",
   CICSBundle = "CICSBundle",
+  CICSTSQueue = "CICSTSQueue",
+  CICSSharedTSQueue = "CICSSharedTSQueue",
 }
 
 export const SupportedResourceTypes: ResourceTypes[] = Object.values(ResourceTypes);