+++ packages/vsce-api/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Zowe Explorer for IBM CICS Transaction Server API wil
 ## Recent Changes
 
 - Add IRemoteFile interface. [#416](github.com/zowe/cics-for-zowe-client/issues/416)
+- Add TS Queue and Shared TS Queue to supported resources. [#415](https://github.com/zowe/cics-for-zowe-client/issues/415)
 
 ## `6.14.0`